Add files via upload
OK so HDV wall amdCity spirit XCP23-4 and city spirit pier1 and hdv wall are the imges hanging and the wall they are on, the other city spirit images are fine art, then the travel photos are examples of wall art, bitht e fine art and wall art will have gifts, and the HDV will be limited to prints and a book. as they are hanging in a promint space in a higer tier hetel in central brighton I am looking at starting price for A3+ at £375 up to A0 for £2250, as the fine arr images take time to create espcially the long exposurte effect images which ar eshot hand held, and then extensive time lining up and layering the images, all done by hand throguth heeditng process, and cull offline images to keep everythgn as sharp as possibl en frame. 
Please guide on pricing but I think that may work OK, maybe A3 at £75 and then up from there only oferring the 4 size options, the trvel images will be prt of the normal wall art, I am looking at strating from 175 for A4 all images arrive frmed and ready to hang, if you could include the option of clack white or natural wood frame as well as with mount and without of r all the prints, and then for gifts I am not sure on the pricing but think the set of 4 mugs witht he city spirit images shoul dbe about £85 to £110, and the individual mugs for 45., tees and hoodies etc I am not surte as they will be mix of fine art designs and my signture logo colloection (which will be just my watermark placed on the arms of sunglasses or my logo on the hoodie, just subtle touches on the xcp signture appareal and notepads, etc.